Moved tests into separat folders. #25
