+++ pim/navigation/__tests__/MainTabNavigator-test.js
