+++ pim/screens/__tests__/GoalScreen-test.js
@@ -1,6 +1,6 @@
 import "react-native";
 import React from "react";
-import Goals from "../screens/GoalsScreen";
+import Goals from "../GoalsScreen";
 import renderer from "react-test-renderer";
 import ShallowRenderer from "react-test-renderer/shallow";
 import MockAsyncStorage from 'mock-async-storage';
