+++ pim/screens/__tests__/NewGoal-test.js
@@ -1,7 +1,7 @@
 import "react-native";
 import React from "react";
 import NewGoal from "../screens/NewGoalScreen";
-import Goals from "../screens/GoalsScreen";
+import Goals from "../GoalsScreen";
 import renderer from "react-test-renderer";
 import ShallowRenderer from "react-test-renderer/shallow";
 //Snapshot testing
